+++ cmake/Android/gradle/upload.gradle.cmake
@@ -48,6 +48,7 @@ publishing {
             artifact("$buildDir/outputs/aar/linphone-sdk-android-debug.aar")
             artifact source: "$buildDir/libs/linphone-sdk-android-sources.jar", classifier: 'sources', extension: 'jar'
             artifact source: "$buildDir/libs/linphone-sdk-android-javadoc.jar", classifier: 'javadoc', extension: 'jar'
+            artifact source: "$buildDir/distributions/linphone-sdk-android-libs-debug.zip", classifier: 'libs-debug', extension: 'zip'
 
             pom {
                 name = 'Linphone'
@@ -73,6 +74,7 @@ publishing {
             artifact("$buildDir/outputs/aar/linphone-sdk-android-release.aar")
             artifact source: "$buildDir/libs/linphone-sdk-android-sources.jar", classifier: 'sources', extension: 'jar'
             artifact source: "$buildDir/libs/linphone-sdk-android-javadoc.jar", classifier: 'javadoc', extension: 'jar'
+            artifact source: "$buildDir/distributions/linphone-sdk-android-libs-debug.zip", classifier: 'libs-debug', extension: 'zip'
 
             pom {
                 name = 'Linphone'