+++ cmake/IOS/linphone-sdk.podspec.cmake
@@ -10,8 +10,8 @@ Pod::Spec.new do |s|
 @LINPHONESDK_LICENSE@LICENSE
     }
   s.author       = { 'Belledonne Communications SARL' => 'linphone-iphone@belledonne-communications.com' }
-  s.platform     = :ios, "12.0"
-  s.source       = { :http => "@LINPHONESDK_IOS_BASE_URL@linphone-sdk-ios-@LINPHONESDK_VERSION@.zip" }
+  s.platform     = :ios, "13.0"
+  s.source       = { :http => "@LINPHONESDK_IOS_BASE_URL@/@LINPHONESDK_NAME@-@LINPHONESDK_VERSION@.zip" }
   s.vendored_frameworks = "@LINPHONESDK_NAME@/apple-darwin/@LINPHONESDK_FRAMEWORK_FOLDER@/**"
   s.pod_target_xcconfig = { 'VALID_ARCHS' => "@VALID_ARCHS@" }
   s.user_target_xcconfig = { 'VALID_ARCHS' => "@VALID_ARCHS@" }